requirements.txt
flask
pandas
scikit-learn
python-dotenv
requests